requirements.txt
Flask==2.0.1
line-bot-sdk==2.2.1
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
